Improve feature Add List
Add property isNewList to setActiveList() to enable auto toggle to edit mode when a new list is added.
Rename onAddItemToList() to addList() to better describe its function.
Remove editNote() as it was a placeholder for a function no longer required.
Add onClickAddList() to handle unique algorithm for Adding a new list.